requirements.txt
Babel==2.12.1
Pillow==9.5.0
tkcalendar==1.6.1
tkcolorpicker==2.1.3
